+++ mne/coreg.py
@@ -42,7 +42,9 @@ head_bem_fname = pformat(bem_fname, name='head')
 fid_fname = pformat(bem_fname, name='fiducials')
 fid_fname_general = os.path.join(bem_dirname, "{head}-fiducials.fif")
 src_fname = os.path.join(bem_dirname, '{subject}-{spacing}-src.fif')
-_head_fnames = (head_bem_fname, pformat(bem_fname, name='head-medium'))
+_head_fnames = (os.path.join(bem_dirname, 'outer_skin.surf'),
+                head_bem_fname,
+                pformat(bem_fname, name='head-medium'))
 _high_res_head_fnames = (os.path.join(bem_dirname, '{subject}-head-dense.fif'),
                          os.path.join(surf_dirname, 'lh.seghead'),
                          os.path.join(surf_dirname, 'lh.smseghead'))