+++ mne/gui/__init__.py
@@ -38,7 +38,8 @@ def combine_kit_markers():
 def coregistration(tabbed=False, split=True, scene_width=None, inst=None,
                    subject=None, subjects_dir=None, guess_mri_subject=None,
                    scene_height=None, head_opacity=None, head_high_res=None,
-                   trans=None, scrollable=True, verbose=None):
+                   trans=None, scrollable=True, project_eeg=False,
+                   verbose=None):
     """Coregister an MRI with a subject's head shape.
 
     The recommended way to use the GUI is through bash with:
@@ -87,6 +88,11 @@ def coregistration(tabbed=False, split=True, scene_width=None, inst=None,
         The transform file to use.
     scrollable : bool
         Make the coregistration panel vertically scrollable (default True).
+    project_eeg : bool
+        If True (default False), project EEG electrodes to nearest
+        head surface vertex.
+
+        .. versionadded:: 0.16
     verbose : bool, str, int, or None
         If not None, override default verbose level (see :func:`mne.verbose`
         and :ref:`Logging documentation <tut_logging>` for more).
@@ -125,7 +131,8 @@ def coregistration(tabbed=False, split=True, scene_width=None, inst=None,
     from ._coreg_gui import CoregFrame, _make_view
     view = _make_view(tabbed, split, scene_width, scene_height, scrollable)
     frame = CoregFrame(inst, subject, subjects_dir, guess_mri_subject,
-                       head_opacity, head_high_res, trans, config)
+                       head_opacity, head_high_res, trans, config,
+                       project_eeg=project_eeg)
     return _initialize_gui(frame, view)